Crate pallet_restricted_tokens

Re-exports

pub use pallet::*;

pub use weights::*;

Modules

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

weights

Autogenerated weights for pallet_restricted_tokens

Structs

FungibleInspectPassthrough

FungiblesInspectPassthrough

TransferDetails

Enums

CurrencyEffects

Represents the trait Currency effects that are called via the pallet-restricted-tokens.

FungibleInspectEffects

Represents the trait fungible::Inspect effects that are called via the pallet-restricted-tokens.

FungibleInspectHoldEffects

Represents the trait fungible::InspectHold effects that are called via the pallet-restricted-tokens.

FungibleMutateEffects

FungibleMutateHoldEffects

Represents the trait fungible::MutateHold effects that are called via the pallet-restricted-tokens.

FungibleTransferEffects

Represents the trait fungible::Transfer effects that are called via the pallet-restricted-tokens.

FungiblesInspectEffects

Represents the trait fungibles::Inspect effects that are called via the pallet-restricted-tokens.

FungiblesInspectHoldEffects

Represents the trait fungibles::InspectHold effects that are called via the pallet-restricted-tokens.

FungiblesMutateEffects

Represents the trait fungibles::Mutate effects that are called via the pallet-restricted-tokens.

FungiblesMutateHoldEffects

Represents the trait fungibles::MutateHold effects that are called via the pallet-restricted-tokens.

FungiblesTransferEffects

Represents the trait fungibles::Transfer effects that are called via the pallet-restricted-tokens.

ReservableCurrencyEffects

Represents the trait Currency effects that are called via the pallet-restricted-tokens.

TokenType