Trait pallet_restricted_tokens::pallet::Config

pub trait Config: Config {

    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type Balance: Parameter + Member + AtLeast32BitUnsigned + Default + Copy + MaybeSerializeDeserialize + MaxEncodedLen + FixedPointOperand;
    type CurrencyId: Parameter + Member + Copy + MaybeSerializeDeserialize + Ord + TypeInfo + MaxEncodedLen;
    type PreExtrTransfer: PreConditions<TransferDetails<Self::AccountId, Self::CurrencyId, Self::Balance>, Result = bool>;
    type PreFungiblesInspect: PreConditions<FungiblesInspectEffects<Self::CurrencyId, Self::AccountId, Self::Balance>, Result = Self::Balance>;
    type PreFungiblesInspectHold: PreConditions<FungiblesInspectHoldEffects<Self::CurrencyId, Self::AccountId, Self::Balance>, Result = bool>;
    type PreFungiblesMutate: PreConditions<FungiblesMutateEffects<Self::CurrencyId, Self::AccountId, Self::Balance>, Result = bool>;
    type PreFungiblesMutateHold: PreConditions<FungiblesMutateHoldEffects<Self::CurrencyId, Self::AccountId, Self::Balance>, Result = bool>;
    type PreFungiblesTransfer: PreConditions<FungiblesTransferEffects<Self::CurrencyId, Self::AccountId, Self::Balance>, Result = bool>;
    type PreFungiblesUnbalanced: PreConditions<FungiblesUnbalancedEffects<Self::CurrencyId, Self::AccountId, Self::Balance>, Result = bool>;
    type Fungibles: Inspect<Self::AccountId, AssetId = Self::CurrencyId, Balance = Self::Balance> + InspectHold<Self::AccountId, Reason = ()> + Mutate<Self::AccountId> + MutateHold<Self::AccountId>;
    type PreCurrency: PreConditions<CurrencyEffects<Self::AccountId, Self::Balance>, Result = bool>;
    type PreReservableCurrency: PreConditions<ReservableCurrencyEffects<Self::AccountId, Self::Balance>, Result = bool>;
    type PreFungibleInspect: PreConditions<FungibleInspectEffects<Self::AccountId, Self::Balance>, Result = Self::Balance>;
    type PreFungibleInspectHold: PreConditions<FungibleInspectHoldEffects<Self::AccountId, Self::Balance>, Result = bool>;
    type PreFungibleMutate: PreConditions<FungibleMutateEffects<Self::AccountId, Self::Balance>, Result = bool>;
    type PreFungibleMutateHold: PreConditions<FungibleMutateHoldEffects<Self::AccountId, Self::Balance>, Result = bool>;
    type PreFungibleTransfer: PreConditions<FungibleTransferEffects<Self::AccountId, Self::Balance>, Result = bool>;
    type NativeFungible: Currency<Self::AccountId, Balance = Self::Balance> + LockableCurrency<Self::AccountId> + ReservableCurrency<Self::AccountId> + Inspect<Self::AccountId, Balance = Self::Balance> + InspectHold<Self::AccountId, Reason = ()> + Mutate<Self::AccountId> + MutateHold<Self::AccountId>;
    type NativeToken: Get<Self::CurrencyId>;
    type WeightInfo: WeightInfo;

}

Configure the pallet by specifying the parameters and types on which it depends.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

Because this pallet emits events, it depends on the runtime’s definition of an event.

type Balance: Parameter + Member + AtLeast32BitUnsigned + Default + Copy + MaybeSerializeDeserialize + MaxEncodedLen + FixedPointOperand

The balance type

type CurrencyId: Parameter + Member + Copy + MaybeSerializeDeserialize + Ord + TypeInfo + MaxEncodedLen

The currency-id of this pallet

type PreExtrTransfer: PreConditions<TransferDetails<Self::AccountId, Self::CurrencyId, Self::Balance>, Result = bool>

Checks the pre conditions for every transfer via the user api (i.e. extrinsics)

type PreFungiblesInspect: PreConditions<FungiblesInspectEffects<Self::CurrencyId, Self::AccountId, Self::Balance>, Result = Self::Balance>

Checks the pre conditions for trait fungibles::Inspect calls

type PreFungiblesInspectHold: PreConditions<FungiblesInspectHoldEffects<Self::CurrencyId, Self::AccountId, Self::Balance>, Result = bool>

Checks the pre conditions for trait fungibles::InspectHold calls

type PreFungiblesMutate: PreConditions<FungiblesMutateEffects<Self::CurrencyId, Self::AccountId, Self::Balance>, Result = bool>

Checks the pre conditions for trait fungibles::Mutate calls

type PreFungiblesMutateHold: PreConditions<FungiblesMutateHoldEffects<Self::CurrencyId, Self::AccountId, Self::Balance>, Result = bool>

Checks the pre conditions for trait fungibles::MutateHold calls

type PreFungiblesTransfer: PreConditions<FungiblesTransferEffects<Self::CurrencyId, Self::AccountId, Self::Balance>, Result = bool>

Checks the pre conditions for trait fungibles::Transfer calls

type PreFungiblesUnbalanced: PreConditions<FungiblesUnbalancedEffects<Self::CurrencyId, Self::AccountId, Self::Balance>, Result = bool>

Checks the pre conditions for trait fungibles::Unbalanced calls

type Fungibles: Inspect<Self::AccountId, AssetId = Self::CurrencyId, Balance = Self::Balance> + InspectHold<Self::AccountId, Reason = ()> + Mutate<Self::AccountId> + MutateHold<Self::AccountId>

type PreCurrency: PreConditions<CurrencyEffects<Self::AccountId, Self::Balance>, Result = bool>

Checks the pre conditions for trait Currency calls

type PreReservableCurrency: PreConditions<ReservableCurrencyEffects<Self::AccountId, Self::Balance>, Result = bool>

Checks the pre conditions for trait ReservableCurrency calls

type PreFungibleInspect: PreConditions<FungibleInspectEffects<Self::AccountId, Self::Balance>, Result = Self::Balance>

Checks the pre conditions for trait fungible::Inspect calls

type PreFungibleInspectHold: PreConditions<FungibleInspectHoldEffects<Self::AccountId, Self::Balance>, Result = bool>

Checks the pre conditions for trait fungible::InspectHold calls

type PreFungibleMutate: PreConditions<FungibleMutateEffects<Self::AccountId, Self::Balance>, Result = bool>

Checks the pre conditions for trait fungible::Mutate calls

type PreFungibleMutateHold: PreConditions<FungibleMutateHoldEffects<Self::AccountId, Self::Balance>, Result = bool>

Checks the pre conditions for trait fungible::MutateHold calls

type PreFungibleTransfer: PreConditions<FungibleTransferEffects<Self::AccountId, Self::Balance>, Result = bool>

Checks the pre conditions for trait fungible::Transfer calls

type NativeFungible: Currency<Self::AccountId, Balance = Self::Balance> + LockableCurrency<Self::AccountId> + ReservableCurrency<Self::AccountId> + Inspect<Self::AccountId, Balance = Self::Balance> + InspectHold<Self::AccountId, Reason = ()> + Mutate<Self::AccountId> + MutateHold<Self::AccountId>

type NativeToken: Get<Self::CurrencyId>

type WeightInfo: WeightInfo